+++ bundle/src/test/java/com/adobe/acs/commons/httpcache/config/impl/RequestParameterHttpCacheConfigExtensionTest.java
@@ -22,10 +22,8 @@ package com.adobe.acs.commons.httpcache.config.impl;
 import com.adobe.acs.commons.httpcache.config.HttpCacheConfig;
 import com.adobe.acs.commons.httpcache.config.impl.keys.KeyValueHttpCacheKey;
 import com.adobe.acs.commons.httpcache.config.impl.keys.helper.KeyValueMapWrapper;
-import com.adobe.acs.commons.httpcache.config.impl.keys.helper.RequestCookieKeyValueWrapperBuilder;
 import com.adobe.acs.commons.httpcache.exception.HttpCacheKeyCreationException;
 import com.adobe.acs.commons.httpcache.exception.HttpCacheRepositoryAccessException;
-import com.google.common.collect.ImmutableSet;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.resource.Resource;
@@ -34,19 +32,22 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.stubbing.Answer;
 
-import javax.servlet.http.Cookie;
 import java.lang.annotation.Annotation;
+import java.util.Collections;
+import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static java.util.Collections.emptyMap;
+import static org.apache.commons.collections.IteratorUtils.asEnumeration;
+import static org.junit.Assert.*;
+import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class RequestCookieCacheExtensionTest {
+public class RequestParameterHttpCacheConfigExtensionTest {
 
     public static final String REQUEST_URI = "/check-in.html";
     private static final String REQUEST_RESOURCE_PATH = "/content/eurowings/de/check-in";
@@ -57,13 +58,10 @@ public class RequestCookieCacheExtensionTest {
     private SlingHttpServletRequest emptyRequest;
 
     @Mock
-    private Cookie presentCookie;
+    private HttpCacheConfig validCacheKeyConfig;
 
     @Mock
-    private Cookie irrelevantCookie;
-
-    private Cookie[] cookies;
-    private Cookie[] emptyCookies;
+    private Resource requestResource;
 
     KeyValueConfig configA = new KeyValueConfig(){
         @Override
@@ -78,7 +76,7 @@ public class RequestCookieCacheExtensionTest {
 
         @Override
         public String[] allowedKeys() {
-            return new String[]{"present-cookie-key", "non-present-cookie"};
+            return new String[]{"present-parameter-key", "non-present-parameter"};
         }
 
         @Override
@@ -105,12 +103,12 @@ public class RequestCookieCacheExtensionTest {
 
         @Override
         public String[] allowedKeys() {
-            return new String[]{"present-cookie-key"};
+            return new String[]{"present-parameter-key"};
         }
 
         @Override
         public String[] allowedValues() {
-            return new String[]{"present-cookie-key=present-cookie-value|value2"};
+            return new String[]{"present-parameter-key=present-parameter-value|value2"};
         }
 
         @Override
@@ -119,33 +117,33 @@ public class RequestCookieCacheExtensionTest {
         }
     };
 
-    @Mock
-    private HttpCacheConfig validCacheKeyConfig;
-
-    @Mock
-    private Resource requestResource;
-
-    private final RequestCookieHttpCacheConfigExtension systemUnderTest = new RequestCookieHttpCacheConfigExtension();
+    private final Map<String,String[]> parameterValueMap = new HashMap<>();
+    private final RequestParameterHttpCacheConfigExtension systemUnderTest = new RequestParameterHttpCacheConfigExtension();
 
 
     @Before
     public void setUp(){
 
-        cookies = new Cookie[]{presentCookie};
-        emptyCookies = new Cookie[]{irrelevantCookie};
+        parameterValueMap.put("present-parameter-key", new String[]{"present-parameter-value"});
 
-        when(emptyRequest.getCookies()).thenReturn(emptyCookies);
-        when(request.getCookies()).thenReturn(cookies);
+        when(request.getParameterNames()).thenAnswer((Answer<Enumeration<String>>) invocationOnMock -> asEnumeration(parameterValueMap.keySet().iterator()));
+        when(emptyRequest.getParameterNames()).thenAnswer((Answer<Enumeration<String>>) invocationOnMock -> asEnumeration(Collections.emptyIterator()));
 
-        when(irrelevantCookie.getName()).thenReturn("bla");
-        when(presentCookie.getName()).thenReturn("present-cookie-key");
-        when(presentCookie.getValue()).thenReturn("present-cookie-value");
+        when(request.getParameterMap()).thenAnswer((Answer<Map<String, String[]>>) invocationOnMock -> parameterValueMap);
+        when(emptyRequest.getParameterMap()).thenAnswer((Answer<Map<String, String[]>>) invocationOnMock -> emptyMap());
+
+
+        when(request.getParameter(anyString())).thenAnswer((Answer<String>) invocationOnMock -> {
+            String key = invocationOnMock.getArguments()[0].toString();
+            return parameterValueMap.get(key)[0];
+        });
 
         when(validCacheKeyConfig.getAuthenticationRequirement()).thenReturn("anonymous");
         when(request.getRequestURI()).thenReturn(REQUEST_URI);
         when(request.getResource()).thenReturn(requestResource);
         when(requestResource.getPath()).thenReturn(REQUEST_RESOURCE_PATH);
     }
+
 
     @Test
     public void test() throws HttpCacheRepositoryAccessException, HttpCacheKeyCreationException {
@@ -157,12 +155,12 @@ public class RequestCookieCacheExtensionTest {
         KeyValueHttpCacheKey cacheKey = (KeyValueHttpCacheKey) systemUnderTest.build(request, validCacheKeyConfig);
         KeyValueMapWrapper map = cacheKey.getKeyValueMap();
 
-        assertTrue(map.containsKey("present-cookie-key"));
-        assertEquals(StringUtils.EMPTY, map.get("present-cookie-key"));
+        assertTrue(map.containsKey("present-parameter-key[0]"));
+        assertEquals(StringUtils.EMPTY, map.get("present-parameter-key[0]"));
 
-        assertFalse(map.containsKey("non-present-cookie"));
+        assertFalse(map.containsKey("non-present-parameter"));
 
-        assertEquals("/check-in.html[CookieKeyValues:present-cookie-key][AUTH_REQ:anonymous]", cacheKey.toString());
+        assertEquals("/check-in.html[RequestParameters:present-parameter-key[0]][AUTH_REQ:anonymous]", cacheKey.toString());
 
     }
 
@@ -175,30 +173,12 @@ public class RequestCookieCacheExtensionTest {
         KeyValueHttpCacheKey cacheKey = (KeyValueHttpCacheKey) systemUnderTest.build(request, validCacheKeyConfig);
         KeyValueMapWrapper map = cacheKey.getKeyValueMap();
 
-        assertTrue(map.containsKey("present-cookie-key"));
-        assertEquals("present-cookie-value", map.get("present-cookie-key"));
+        assertTrue(map.containsKey("present-parameter-key[0]"));
+        assertEquals("present-parameter-value", map.get("present-parameter-key[0]"));
 
-        assertFalse(map.containsKey("non-present-cookie"));
+        assertFalse(map.containsKey("non-present-parameter"));
 
-        assertEquals("/check-in.html[CookieKeyValues:present-cookie-key=present-cookie-value][AUTH_REQ:anonymous]", cacheKey.toString());
+        assertEquals("/check-in.html[RequestParameters:present-parameter-key[0]=present-parameter-value][AUTH_REQ:anonymous]", cacheKey.toString());
 
     }
-
-    @Test
-    public void testKey() throws HttpCacheRepositoryAccessException, HttpCacheKeyCreationException {
-        systemUnderTest.activate(configA);
-        assertTrue(systemUnderTest.accepts(request, null));
-        assertFalse(systemUnderTest.accepts(emptyRequest, null));
-
-        ImmutableSet<String> allowedKeys = ImmutableSet.of("present-cookie-key", "non-present-cookie-key");
-        ImmutableSet<Cookie> presentValues = ImmutableSet.of(presentCookie);
-
-        Map<String, String> cookieKeyValues = new HashMap<>();
-        KeyValueMapWrapper requestCookieKeyValueMap = new RequestCookieKeyValueWrapperBuilder(allowedKeys, cookieKeyValues, presentValues).build();
-
-        KeyValueHttpCacheKey cookieCacheKey = new KeyValueHttpCacheKey(REQUEST_URI, validCacheKeyConfig, requestCookieKeyValueMap);
-
-        assertTrue(systemUnderTest.doesKeyMatchConfig(cookieCacheKey, validCacheKeyConfig));
-    }
-
 }